cambiando codigo del ej3 para que funcione correctamente y pase los tests. Cambiando los tests del ej07 y 08 para que sean correctos

+++ src/tp03/ej03.rs
@@ -71,7 +71,7 @@ impl Fecha {
     }
     pub fn restar_dias(&mut self, mut dias: u32) {
         while dias > 0 {
-            if (self.dia - dias) < 1 {
+            if (self.dia as i32 - dias as i32) < 1 {
                 if self.mes == 1 {
                     self.mes = 12;
                     self.anio -= 1;
